tests/requirements.txt
pyTest
pyYAML
PsycoPG2-binary
Pandas
